+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#155](https://github.com/os2display/display-templates/pull/155)
+  - Added single-booking layout to calendar template.
+  - Removed defaultProps.
+
 ## [2.0.1] - 2024-08-14
 
 - [#154](https://github.com/os2display/display-templates/pull/154)
@@ -20,11 +24,11 @@ All notable changes to this project will be documented in this file.
 - [#148](https://github.com/os2display/display-templates/pull/148)
   - Changed theme load following API change.
   - Fixed path issue in #141.
-- [#141](https://github.com/os2display/display-templates/pull/141)
-  - Improve build for local development
 - [#146](https://github.com/os2display/display-templates/pull/146)
   - Fixed issue where image-text image cycle is not restarting when only on slide is displayed.
   - Fixed issue with slideshow default duration being 5000 s.
+- [#141](https://github.com/os2display/display-templates/pull/141)
+  - Improved build for local development
 
 ## [1.3.1] - 2023-10-26